move catalyst.passes.pipeline to top level, i.e. catalyst.pipeline
