+++ frontend/catalyst/__init__.py
@@ -89,6 +89,7 @@ from catalyst.autograph import __all__ as _autograph_functions
 from catalyst.compiler import CompileOptions
 from catalyst.debug.assertion import debug_assert
 from catalyst.jit import QJIT, qjit
+from catalyst.passes import pipeline
 from catalyst.utils.exceptions import (
     AutoGraphError,
     CompileError,
@@ -186,6 +187,7 @@ __all__ = (
     "debug_assert",
     "CompileOptions",
     "debug",
+    "pipeline",
     *_api_extension_list,
     *_autograph_functions,
 )
