+++ frontend/test/lit/test_peephole_optimizations.py
@@ -29,9 +29,9 @@ import shutil
 import pennylane as qml
 from lit_util_printers import print_jaxpr, print_mlir
 
-from catalyst import qjit
+from catalyst import qjit, pipeline
 from catalyst.debug import get_compilation_stage
-from catalyst.passes import cancel_inverses, merge_rotations, pipeline
+from catalyst.passes import cancel_inverses, merge_rotations
 
 
 def flush_peephole_opted_mlir_to_iostream(QJIT):
